Original change: https://android-review.googlesource.com/c/platform/test/dittosuite/+/1831953

+++ src/result.cpp
@@ -38,7 +38,7 @@ namespace dittosuite {
 
 Result::Result(const std::string& name, const int& repeat) : name_(name), repeat_(repeat) {}
 
-void Result::AddMeasurement(const std::string& name, const std::vector<int64_t>& samples) {
+void Result::AddMeasurement(const std::string& name, const std::vector<double>& samples) {
   samples_[name] = samples;
   AnalyseMeasurement(name);
 }
@@ -47,7 +47,7 @@ void Result::AddSubResult(std::unique_ptr<Result> result) {
   sub_results_.push_back(std::move(result));
 }
 
-std::vector<int64_t> Result::GetSamples(const std::string& measurement_name) const {
+std::vector<double> Result::GetSamples(const std::string& measurement_name) const {
   return samples_.find(measurement_name)->second;
 }
